Move map generation code into its own namespace
